requirements.txt
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
beautifulsoup4==4.9.3
boto3==1.17.86
botocore==1.20.86
colorama==0.4.4
decorators==2.0.7
Django==3.2.3
django-bootstrap-v5==1.0.4
django-storages==1.11.1
django-taggit==2.1.0
django-tinymce==3.4.0
gunicorn==20.1.0
iniconfig==1.1.1
jmespath==0.10.0
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5
whitenoise==5.2.0
